+++ core/src/main/java/com/graphhopper/routing/util/parsers/OSMMaxSpeedParser.java
@@ -21,7 +21,6 @@ import com.graphhopper.reader.ReaderWay;
 import com.graphhopper.routing.ev.*;
 import com.graphhopper.routing.util.countryrules.CountryRule;
 import com.graphhopper.routing.util.parsers.helpers.OSMValueExtractor;
-import com.graphhopper.routing.util.spatialrules.SpatialRuleSet;
 import com.graphhopper.routing.util.TransportationMode;
 import com.graphhopper.storage.IntsRef;